requirements.txt
Django~=4.0.2
djangorestframework~=3.13.1
Pillow~=9.0.1
django-phonenumber-field~=6.1.0
drf-yasg
djangorestframework-simplejwt
psycopg2